Lots of simplification and refactoring.
In particular, moving closer to Entry being a higher-level object that does
demand-loading and exposes a more useful API for formatting.

Also removed the schema migration stuff because in the end all files get
scanned on startup anyway so it's not like it's any faster to migrate a
schema than to just rebuild the whole freaking database.