Allow caching to actually, y'know, cache.
Back in the day, Publ cached at the request level, which meant that response codes would be cached too, which was of course bad; the fix at the time was to not cache responses that could result in a 304 NOT MODIFIED. Since then (mostly to add authorization) things were refactored to put the caching at the render level, but the aforementioned cache elision was never removed. Oops.