Improve template caching
This prevents a few situations where a template has been
moved/deleted/edited in a persistent cache scenario.

If a template was removed, it would have caused errors until the old
template mapping expired from the cache. This prevents that from happening.

If a template was modified, this now causes the cache lookup against the
old template to fail, improving the responsiveness for site updates.