Fix the lack of image cache purging
Fixes #456

The cache purge mechanism was running in the image threadpool, but wasn't being wrapped in the app context, which was necessary to obtain the configuration information. This fix simply has the in-context wrapper compute the necessary information.

A better fix would be to make maintenance tasks run in their own threadpool; see #457