Ignore exceptions in getting the user if inside an exception handler
If a 404 error occurs when there's also an invalid Authorization header,
this can cause a cascade of exceptions because the error handler tries to parse
the Authorization header while it's already handling an exception. This change
prevents this from occurring, by allowing `user.get_active()` to continue with
the null user if getting the user identifier fails.

Fixes #581